# app/static/swagger.yaml
openapi: 3.0.0
info:
  title: Bus Payments & Rewards API
  description: Comprehensive API for bus payment system with offers, rewards, bookings, and refunds
  version: 1.0.0
servers:
  - url: http://localhost:{port}
    description: Local development server
    variables:
      port:
        default: '5006'
        description: Server port number
components:
  schemas:
...
    CreditCardProduct:
      type: string
      enum:
        - BASIC
        - SILVER
        - GOLD
        - PLATINUM
...
    RewardStatus:
      type: string
      enum:
        - EARNED
        - REDEEMED
        - EXPIRED
...
paths:
...
  /api/customers:
    get:
      tags:
        - Customers
      summary: List all customers
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: per_page
          schema:
            type: integer
            default: 10
        - in: query
          name: email
          schema:
            type: string
      responses:
        '200':
          description: List of customers
          content:
            application/json:
              schema:
                type: object
                properties:
                  customers:
                    type: array
                    items:
                      type: object
                  total:
                    type: integer
                  pages:
                    type: integer
                  current_page:
                    type: integer
                  per_page:
                    type: integer
    post:
      tags:
        - Customers
      summary: Create new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                first_name:
                  type: string
                last_name:
                  type: string
                email:
                  type: string
                  format: email
                phone:
                  type: string
                date_of_birth:
                  type: string
                  format: date
                address:
                  type: string
              required:
                - first_name
                - last_name
                - email
      responses:
        '201':
          description: Customer created successfully
        '400':
          description: Invalid data
        '409':
          description: Email already exists
  /api/customers/{customer_id}:
    get:
      tags:
        - Customers
      summary: Get customer details
      parameters:
        - in: path
          name: customer_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Customer details
        '404':
          description: Customer not found
    put:
      tags:
        - Customers
      summary: Update customer
      parameters:
        - in: path
          name: customer_id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                first_name:
                  type: string
                last_name:
                  type: string
                email:
                  type: string
                  format: email
                phone:
                  type: string
                date_of_birth:
                  type: string
                  format: date
                address:
                  type: string
      responses:
        '200':
          description: Customer updated successfully
        '400':
          description: Invalid data
        '404':
          description: Customer not found
    delete:
      tags:
        - Customers
      summary: Delete customer
      parameters:
        - in: path
          name: customer_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Customer deleted successfully
        '404':
          description: Customer not found
  /api/customers/{customer_id}/credit-cards:
    get:
      tags:
        - Credit Cards
      summary: Get customer credit cards
      parameters:
        - in: path
          name: customer_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of customer credit cards
        '404':
          description: Customer not found
    post:
      tags:
        - Credit Cards
      summary: Add credit card to customer
      parameters:
        - in: path
          name: customer_id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                card_number:
                  type: string
                card_holder_name:
                  type: string
                expiry_month:
                  type: integer
                  minimum: 1
                  maximum: 12
                expiry_year:
                  type: integer
                product_type:
                  $ref: '#/components/schemas/CreditCardProduct'
                credit_limit:
                  type: number
                  format: float
                  minimum: 0
                is_active:
                  type: boolean
                  default: true
              required:
                - card_number
                - card_holder_name
                - expiry_month
                - expiry_year
                - product_type
                - credit_limit
      responses:
        '201':
          description: Credit card added successfully
        '400':
          description: Invalid data
        '409':
          description: Card number already exists
...
  /api/payments/customer/{customer_id}:
    get:
      tags:
        - Payments
      summary: Get customer payments
      parameters:
        - in: path
          name: customer_id
          required: true
          schema:
            type: integer
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: per_page
          schema:
            type: integer
        - in: query
          name: start_date
          schema:
            type: string
            format: date
        - in: query
          name: end_date
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Customer payment history
...
  /api/tokens/customer/{customer_id}:
    get:
      tags:
        - Card Tokens
      summary: Get customer tokens
      parameters:
        - in: path
          name: customer_id
          required: true
          schema:
            type: integer
        - in: query
          name: is_active
          schema:
            type: boolean
      responses:
        '200':
          description: List of customer tokens
        '404':
          description: Customer not found
        '500':
          description: Internal server error
...
  /api/rewards/customer/{customer_id}:
    get:
      tags:
        - Rewards
      summary: Get customer rewards
      parameters:
        - in: path
          name: customer_id
          required: true
          schema:
            type: integer
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: per_page
          schema:
            type: integer
            default: 10
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/RewardStatus'
        - in: query
          name: start_date
          schema:
            type: string
            format: date-time
        - in: query
          name: end_date
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: List of customer rewards
        '400':
          description: Invalid parameters
        '404':
          description: Customer not found
        '500':
          description: Internal server error
...